@@ -1,17 +1,13 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
 
+import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
-import { NotificationModule } from './notification/notification.module';
-import { SchedulerModule } from './scheduler/scheduler.module';
 import { NoContentComponent } from './no-content/no-content.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { SubjectsModule } from './subjects/subjects.module';
-
 
 @NgModule({
   declarations: [
@@ -22,9 +18,6 @@ import { SubjectsModule } from './subjects/subjects.module';
     BrowserModule,
     BrowserAnimationsModule,
     SharedModule.forRoot(),
-    NotificationModule,
-    SchedulerModule,
-    SubjectsModule,
     FlexLayoutModule,
     AppRoutingModule
   ],
