@@ -4,6 +4,7 @@ import { SchedulerService } from './scheduler.service';
 import { SubjectCourseListComponent } from './subject-course-list.component';
 import { SharedModule } from '../shared/shared.module';
 import { SchedulerRoutingModule } from './scheduler-routing.module';
+import { SchedulerEditComponent } from './scheduler-edit.component';
 
 @NgModule({
   imports: [
@@ -11,7 +12,8 @@ import { SchedulerRoutingModule } from './scheduler-routing.module';
     SharedModule,
     SchedulerRoutingModule,
   ],
-  declarations: [SubjectCourseListComponent],
+  declarations: [SubjectCourseListComponent,
+    SchedulerEditComponent],
   providers: [SchedulerService]
 })
 export class SchedulerModule {
