@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'ek-subjects-edit',
@@ -6,10 +7,17 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./subjects-edit.component.scss']
 })
 export class SubjectsEditComponent implements OnInit {
+  public subjectForm: FormGroup;
 
-  constructor() { }
+  constructor(public fb: FormBuilder) {
+  }
 
   ngOnInit() {
   }
 
+  private formBuild(): void {
+    this.subjectForm = this.fb.group({
+      persistenceDuration: ['']
+    }, {updateOn: 'blur'});
+  }
 }
